requirements.txt
pyyaml
python-gitlab
gitlab
